+++ src/types/internal.ts
@@ -16,7 +16,7 @@ export type Uint8_t = number & { __type: 'uint8_t' }
 
 // Reexport blockchain spec
 export { AddressType, CertificateType, RelayType }
-export { GetVersionResponse as Version } from './public'
+export { Version } from './public'
 // Our types
 export const KEY_HASH_LENGTH = 28;
 export const TX_HASH_LENGTH = 32;